constant added to db fields extended edit form implementation started some clean up

+++ src/main/java/com/lv/reg/dao/CustomerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Collection;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
     Collection<Customer> findAllByEmail(String email);
 }
